+++ src/Livewire/Organizations/OrganizationShow.php
@@ -2,19 +2,22 @@
 
 namespace VentureDrake\LaravelCrm\Livewire\Organizations;
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 use Mary\Traits\Toast;
 use VentureDrake\LaravelCrm\Models\Organization;
 
 class OrganizationShow extends Component
 {
-    use Toast;
+    use AuthorizesRequests, Toast;
 
     public Organization $organization;
 
     public function delete($id)
     {
         if ($organization = Organization::find($id)) {
+            $this->authorize('delete', $organization);
+
             $organization->delete();
 
             $this->success(ucfirst(trans('laravel-crm::lang.organization_deleted')), redirectTo: route('laravel-crm.organizations.index'));
