+++ src/Livewire/People/PersonShow.php
@@ -2,19 +2,22 @@
 
 namespace VentureDrake\LaravelCrm\Livewire\People;
 
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Livewire\Component;
 use Mary\Traits\Toast;
 use VentureDrake\LaravelCrm\Models\Person;
 
 class PersonShow extends Component
 {
-    use Toast;
+    use AuthorizesRequests, Toast;
 
     public Person $person;
 
     public function delete($id)
     {
         if ($person = Person::find($id)) {
+            $this->authorize('delete', $person);
+
             $person->delete();
 
             $this->success(ucfirst(trans('laravel-crm::lang.person_deleted')), redirectTo: route('laravel-crm.people.index'));
